requirements.txt
snowflake-connector-python
snowflake-connector-python[pandas]
altair
streamlit~=1.10.0
seaborn
matplotlib